+++ test/integration/bigip/property/propertiesTlsClient.js
@@ -218,6 +218,11 @@ describe('TLS_Client', function () {
                 name: 'sslSignHash',
                 inputValue: [undefined, 'sha256', undefined],
                 expectedValue: ['any', 'sha256', 'any']
+            },
+            {
+                name: 'handshakeTimeout',
+                inputValue: [undefined, 1234, undefined],
+                expectedValue: [10, 1234, 10]
             }
         ];